requirements.txt
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
beautifulsoup4==4.8.0
certifi==2019.3.9
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.5
cssselect==1.0.3
future==0.17.1
httplib2==0.12.3
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
lxml==4.3.1
mysqlclient==1.4.2.post1
oauth2==1.9.0.post1
oauthlib==3.0.1
parsel==1.5.1
Pillow==5.4.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
PyMySQL==0.9.3
pyOpenSSL==19.0.0
python-scrapyd-api==2.1.2
python-twitter==3.5
pytz==2019.1
pywin32==224
queuelib==1.5.0
requests==2.21.0
requests-oauthlib==1.2.0
Scrapy==1.6.0
scrapyd==1.2.1
scrapyd-client==1.1.0
service-identity==18.1.0
simplejson==3.16.0
six==1.12.0
soupsieve==1.9.2
Twisted==18.9.0
urllib3==1.24.3
w3lib==1.20.0
zope.interface==4.6.0
